Optimized vault with local file caching

+++ Parallel.Core/IO/Syncing/BaseSyncManager.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright 2026 Kyle Ebbinga
 
+using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
 using Parallel.Core.Database;
 using Parallel.Core.Database.Contexts;
@@ -46,9 +47,15 @@ namespace Parallel.Core.IO.Syncing
         }
 
         /// <inheritdoc />
-        public async Task<bool> ConnectAsync()
+        public async Task<bool> ConnectAsync(bool force = false)
         {
-            Log.Debug("[{LocalVaultId}] Connecting...", LocalVault.Id);
+            if (!await StorageProvider.CheckConnectionAsync())
+            {
+                Log.Error("[{LocalVaultId}] Failed to connect to vault!", LocalVault.Id);
+                return false;
+            }
+            
+            
             string root = PathBuilder.GetRootDirectory(LocalVault);
             if (!await StorageProvider.ExistsAsync(root))
             {
@@ -56,43 +63,50 @@ namespace Parallel.Core.IO.Syncing
                 Log.Debug("Created root directory: {Root}", root);
             }
             
+            // Checks temp files for a local download of the config file
+            if (!File.Exists(TempConfigFile) || File.GetLastWriteTimeUtc(TempConfigFile) <= DateTime.UtcNow.AddHours(-6) || force)
+            {
+                Log.Debug($"No local config file found. Fetching...");
+                if (!await StorageProvider.ExistsAsync(PathBuilder.GetConfigurationFile(LocalVault)))
+                {
+                    RemoteVault = new RemoteVaultConfig(LocalVault);
+                    RemoteVault.IgnoreDirectories.Add(PathBuilder.GetRootDirectory(LocalVault));
+                    RemoteVault.Save(TempConfigFile);
+
+                    Log.Debug("Created config file: {TempConfigFile}", TempConfigFile);
+                }
+                else
+                {
+                    await StorageProvider.DownloadFileAsync(new LocalFile(TempConfigFile), PathBuilder.GetConfigurationFile(LocalVault));
+                    Log.Debug("Downloaded file: {TempConfigFile}", TempConfigFile);
+                }
+            }
+
+            // Checks temp files for a local download of the database file
+            if (!File.Exists(TempDbFile) || File.GetLastWriteTimeUtc(TempDbFile) <= DateTime.UtcNow.AddHours(-3) || force)
+            {
+                Log.Debug($"No local database file found. Fetching...");
+                if (!await StorageProvider.ExistsAsync(PathBuilder.GetDatabaseFile(LocalVault)))
+                {
+                    Log.Debug("Create db file: {TempDbFile}", TempDbFile);
+                    if (File.Exists(TempDbFile)) File.Delete(TempDbFile);
+                    Database = new SqliteContext(TempDbFile);
+                    await Database.InitializeAsync();
+                }
+                else
+                {
+                    string remoteDbFile = PathBuilder.GetDatabaseFile(LocalVault);
+                    await StorageProvider.DownloadFileAsync(new LocalFile(TempDbFile), remoteDbFile);
+                    Log.Debug("Downloaded file: {TempDbFile}", TempDbFile);
+                }   
+            }
             
-
-            if (!await StorageProvider.ExistsAsync(PathBuilder.GetConfigurationFile(LocalVault)))
-            {
-                RemoteVault = new RemoteVaultConfig(LocalVault);
-                RemoteVault.IgnoreDirectories.Add(PathBuilder.GetRootDirectory(LocalVault));
-                RemoteVault.Save(TempConfigFile);
-
-                Log.Debug("Created config file: {TempConfigFile}", TempConfigFile);
-            }
-            else
-            {
-                await StorageProvider.DownloadFileAsync(new LocalFile(TempConfigFile), PathBuilder.GetConfigurationFile(LocalVault));
-                RemoteVaultConfig? config = RemoteVaultConfig.Load(TempConfigFile);
-                if (config == null) return false;
-                RemoteVault = config;
-                LocalVault.Name = config.Name;
-                Log.Debug("Downloaded file: {TempConfigFile}", TempConfigFile);
-            }
-
-            if (!await StorageProvider.ExistsAsync(PathBuilder.GetDatabaseFile(LocalVault)))
-            {
-                if (File.Exists(TempDbFile)) File.Delete(TempDbFile);
-                Database = new SqliteContext(TempDbFile);
-                await Database.InitializeAsync();
-
-                Log.Debug("Create db file: {TempDbFile}", TempDbFile);
-            }
-            else
-            {
-                string remoteDbFile = PathBuilder.GetDatabaseFile(LocalVault);
-                //await StorageProvider.CloneFileAsync(remoteDbFile, remoteDbFile + ".old");
-                await StorageProvider.DownloadFileAsync(new LocalFile(TempDbFile), remoteDbFile);
-                Database = new SqliteContext(TempDbFile);
-
-                Log.Debug("Downloaded file: {TempDbFile}", TempDbFile);
-            }
+            // Load the temp files
+            Database = new SqliteContext(TempDbFile);
+            RemoteVaultConfig? config = RemoteVaultConfig.Load(TempConfigFile);
+            if (config == null) return false;
+            RemoteVault = config;
+            LocalVault.Name = config.Name;
 
             Log.Information("[{LocalVaultId}] Connected", LocalVault.Id);
             return true;
